feat: drop all variables except _price and _serialized_ledger
